refactor(jolt-sumcheck): rename batched.rs to batched_verifier.rs
The module holds only the verifier type, so match the existing
`verifier.rs` naming pattern rather than the stubby `batched`.
