+++ crates/jolt-sumcheck/src/batched_verifier.rs
